feat: 체스 기물에 색상 추가
