--- src/main/java/chess/domain/Bishop.java
@@ -1,8 +1,0 @@
-package chess.domain;
-
-public class Bishop extends Piece {
-
-    protected Bishop(final int rank, final int file) {
-        super(rank, file);
-    }
-}
