--- src/main/java/chess/domain/King.java
@@ -1,8 +1,0 @@
-package chess.domain;
-
-public class King extends Piece {
-
-    protected King(final int rank, final int file) {
-        super(rank, file);
-    }
-}
