--- src/main/java/chess/domain/Knight.java
@@ -1,8 +1,0 @@
-package chess.domain;
-
-public class Knight extends Piece {
-
-    protected Knight(final int rank, final int file) {
-        super(rank, file);
-    }
-}
