--- src/main/java/chess/domain/Pawn.java
@@ -1,8 +1,0 @@
-package chess.domain;
-
-public class Pawn extends Piece {
-
-    protected Pawn(final int rank, final int file) {
-        super(rank, file);
-    }
-}
