--- src/main/java/chess/domain/Queen.java
@@ -1,8 +1,0 @@
-package chess.domain;
-
-public class Queen extends Piece{
-
-    protected Queen(final int rank, final int file) {
-        super(rank, file);
-    }
-}
