--- src/main/java/chess/domain/Rook.java
@@ -1,8 +1,0 @@
-package chess.domain;
-
-public class Rook extends Piece {
-
-    protected Rook(final int rank, final int file) {
-        super(rank, file);
-    }
-}
