+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,0 +1,8 @@
+package chess.domain.piece;
+
+public class Bishop extends Piece {
+
+    protected Bishop(final int rank, final int file, final Color color) {
+        super(rank, file, color);
+    }
+}
