+++ src/main/java/chess/domain/piece/Color.java
@@ -1,0 +1,6 @@
+package chess.domain.piece;
+
+public enum Color {
+    WHITE,
+    BLACK;
+}
