+++ src/main/java/chess/domain/piece/King.java
@@ -1,0 +1,8 @@
+package chess.domain.piece;
+
+public class King extends Piece {
+
+    protected King(final int rank, final int file, final Color color) {
+        super(rank, file, color);
+    }
+}
