+++ src/main/java/chess/domain/piece/Knight.java
@@ -1,0 +1,8 @@
+package chess.domain.piece;
+
+public class Knight extends Piece {
+
+    protected Knight(final int rank, final int file, final Color color) {
+        super(rank, file, color);
+    }
+}
