+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,0 +1,8 @@
+package chess.domain.piece;
+
+public class Pawn extends Piece {
+
+    protected Pawn(final int rank, final int file, final Color color) {
+        super(rank, file, color);
+    }
+}
