+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,0 +1,8 @@
+package chess.domain.piece;
+
+public class Queen extends Piece {
+
+    protected Queen(final int rank, final int file, final Color color) {
+        super(rank, file, color);
+    }
+}
