+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,0 +1,8 @@
+package chess.domain.piece;
+
+public class Rook extends Piece {
+
+    protected Rook(final int rank, final int file, final Color color) {
+        super(rank, file, color);
+    }
+}
