+++ src/test/java/chess/domain/piece/PieceTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -13,9 +13,9 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 class PieceTest {
 
     @Test
-    void 위치를_가지고_생성된다() {
+    void 위치와_색상을_가지고_생성된다() {
         // when & then
-        assertDoesNotThrow(() -> new Piece(1, 1) {
+        assertDoesNotThrow(() -> new Piece(1, 1, Color.WHITE) {
         });
     }
 }
